+++ playwright.config.ts
@@ -17,7 +17,7 @@ export default defineConfig({
     /* Fail the build on CI if you accidentally left test.only in the source code. */
     forbidOnly: !!process.env.CI,
     /** GitHub currently permits 2 workers, so we use 2. */
-    workers: process.env.CI ? 2 : undefined,
+    workers: process.env.CI ? 1 : 1,
     /* Retry once on CI, never locally */
     retries: process.env.CI ? 1 : 0,
     /* Reporter to use. See https://playwright.dev/docs/test-reporters */